Move ActionsRepository to actions package
