+++ domain/src/main/java/com/aconno/sensorics/domain/interactor/ifttt/action/UpdateActionUseCase.kt
@@ -1,7 +1,7 @@
 package com.aconno.sensorics.domain.interactor.ifttt.action
 
 import com.aconno.sensorics.domain.actions.Action
-import com.aconno.sensorics.domain.ifttt.ActionsRepository
+import com.aconno.sensorics.domain.actions.ActionsRepository
 import com.aconno.sensorics.domain.interactor.type.CompletableUseCaseWithParameter
 import io.reactivex.Completable
